Add files via upload
make sure the protobuf is up to date. it gets downgraded with the keras-cv install.
$ pip install keras-cv
then protobuf will be downgraded. in my case it was 4.21.12, 
so run the code:
$ pip install --upgrade protobuf
to make sure it is 6.33.0

you can ignore this message if it is prompted:
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
tensorflow-metadata 1.17.2 requires protobuf<4.22,>=4.21.6; python_version < "3.11", but you have protobuf 6.33.0 which is incompatible.


otherwise no issues.